Fixed time not appearing for insta-connects/host errors, and reuse existing flash policy file
I keep saying final, eventually it'll be true!
Fixed the time not appearing if clients are trying to connect before timer first ticks, or if there's a host error. Timer tick is what updates the time used in console.
Allows flash policy file to be used (and not deleted on app close) if it already exists on app startup.