Fixed non-capped server builds
Fixed the clientstats constructor used when there is no cap (and no stats related to it).
Also a minor change to checking whether prort was fixed.
Set the server back to uncapped for the repo. I may have to create project definitions for the capped versions later, since this is easily and repeatedly missed.
(Cap increased to 500kb total, but commented out.)