Update Bluewing Server to build 19
Build 19 has no major changes, but has a couple rare-case fixes, as well as some increased documentation in the ping function. It's also a rebuild for the steady_clock issue.
Changed the error message created by Lacewing, so there's not an awkward line break.
The error of clients remaining connected constantly turned out to be an error in the standard library provided in Visual Studio 2019, which I've documented more in the ansi-build-18 release.
In preparation for release, now server messages also count as activity - if server has a handler set up for them. No server-message handler = not activity.