Cleanup of Linux and README
The Linux project would interfere with Windows' output and intermediate files, so moved its output folders so it won't do that.
README file had no extension, which meant it slipped past Git and EditorConfig settings on what line endings and encodings to use. It was easier to give it an extension than remake both gitattributes and editorconfig files.