Fix IP addresses printing on Unix
The IP-to-text functions were producing different formats, resulting in lw_addr_pretty_string() aborting with "".
Standardized them to Windows, partially because they're actually following the standard, and partially because I won't have to patch lw_addr_pretty_string().
It wouldn't surprise me if inet_ntop()'s output varied in different flavours and versions of Unix-based OSes, but hopefully not. Cross that when we come to it.