Removed unused SPDY code/files; SSL define fix
SPDY was deprecated in favour of HTTP/2 in all major browsers in 2021.
I doubt a Blue server, a WebSocket-only HTTP server that only serves a single html page to browsers, will have use of HTTP/2 anyway, as it's a multiple-response-file system.

Also moved the deps folder out of src folder in the vcxproj files, since it wasn't in there on filesystem.

ENABLE_SSL was undefined on Unix config header; re-defined it and removed from vcxproj.
There was one server.c section it was not define-guarding; added it there.
Android and iOS do not come with system config files for openssl - so I will have to gather the correct openssl files and libraries for them. Fun.