Clean up of Windows and SSL files
Filled in all the liblacewing SSL error reporting holes.
Added error number when loading the SSL certificate files fails.
Fixed a Debug-only crash when reporting error in writing to socket, due to missing argument.

I'm keeping the client SSL files so they're not absent should someone use the Lacewing folder as a base for their client SSL, or should I extend Lacewing clients later for SSL support.
But some SSL files aren't used at all, or weren't referenced in the vcxproj file correctly.

Removed the webstate enum - it wasn't used. I was planning a multi-stage switch over, so the raw socket port could accept websocket clients, but with webserver taking the HTTP side and upgrade, all that fluff with webstate wasn't necessary. In theory, the ports can be merged, if you add a layer between to send raw to Lacewing or to HTTP parsers - in practice, that's ignoring the point of ports, that protocols become apparent by responses to their port.