Windows: Fix WebSocket secure shared messages
The encryption of WebSocket secure would be applied to the original message. This was bad news if it was a message that was going to be forwarded to clients after them in the channel's client list, as it meant corrupted messages.

This reflects  SortaCore/MMF2Exts@982ec387c2a072180a7db12ac27e1aed7f6ab208.