Minify chances of crashing with new action queue
This does not really apply to bluewing-cpp-server as it's a single-threaded server. However, it is a significant edit.

The main thread now queues things for the lacewing event pump to processs, if event pump is not being ticked in main thread. Only unhosting is a blocking action that freezes the main thread until action is processed; everything else is queued and processed later.
This is now build 38 as MMF2Exts would have it (specifically, as SortaCore/MMF2Exts@d603e2942011c3505969f80fd84117296dfe6efc would have it).

A few improvements will still happen to bluewing-cpp-server before build 38 is released here.